+++ woocommerce/emails/customer-completed-order.php
@@ -47,10 +47,14 @@ do_action( 'woocommerce_email_order_meta', $order, $sent_to_admin, $plain_text, 
  * @hooked WC_Emails::customer_details() Shows customer details
  * @hooked WC_Emails::email_address() Shows email address
  */
-do_action( 'woocommerce_email_customer_details', $order, $sent_to_admin, $plain_text, $email );
+do_action( 'woocommerce_email_customer_details', $order, $sent_to_admin, $plain_text, $email ); 
+
+/*<p>
+<?php esc_html_e( 'Thanks for shopping with us.', 'woocommerce' ); ?>
+</p>*/
 
 /**
- * Custom Code to add Amazon Locals button
+ * Custom Code to add a promotional button
  * For Button Poetry
  */
 ?><table width="100%" border="0" cellspacing="0" cellpadding="0">
@@ -59,13 +63,15 @@ do_action( 'woocommerce_email_customer_details', $order, $sent_to_admin, $plain_
 			<center>
 				<table border="0" cellspacing="0" cellpadding="0" style="display:inline-block;margin-left:auto;margin-right:auto;">
 					<tr>
-						<td align="center" style="border-radius: 3px;" bgcolor="#e9703e">
-							<a href="https://www.amazon.com/shop/button" target="_blank" style="font-size: 16px; color: #ffffff; text-decoration: none; text-decoration: none;border-radius: 3px; padding: 8px 14px; border: 1px solid #e9703e; display: inline-block;">Button
-								Poetry on Amazon Local →</a></td>
+						<td align="center" style="padding-bottom:10px;font-size:1.1em"><strong>Love Poetry Videos?</strong> Try <a href="https://tv.buttonpoetry.com/">Button TV</a> on us! Get<br><em>3 months FREE</em> with discount code <strong>TRYBTV</strong></td>
+					</tr>
+					<tr>
+						<td align="center" style="border-radius: 3px;" bgcolor="#e9703e">                            
+							<a href="https://tv.buttonpoetry.com/join" target="_blank" style="font-size: 16px; color: #ffffff; text-decoration: none; text-decoration: none;border-radius: 3px; padding: 8px 14px; border: 1px solid #e9703e; display: inline-block;">Watch Button TV Now! →</a></td>
 					</tr>
 					<tr>
 							<td align="center" style="font-size:14px;padding-top:6px;">
-								<em>Explore Button's recommended books and audio on Amazon!</em>
+								<em>Coupon valid on Button TV Monthly Subscription</em>
 							</td>
 						</tr>
 				</table>
@@ -74,9 +80,6 @@ do_action( 'woocommerce_email_customer_details', $order, $sent_to_admin, $plain_
 	</tr>
 </table>
 
-<p>
-<?php esc_html_e( 'Thanks for shopping with us.', 'woocommerce' ); ?>
-</p>
 <?php
 
 /*